#43  Move SelectStatement from cql.ast to dml package
